Makefile: clang-14: Disable 'unused-but-set-variable' for now
This warning is created when a variable is set but not used due to certain
reasons like it might be ifdef-ed with some condition or it might be a
part of a config or it be a debug config.

Anyway, we can hotfix by 'maybe_unused' macro but again we need to edit a
hell lot of drivers. Since this warning is harmless, we can just
suppress the warning as of now.
Anytime later in our free-time, we can just fix the warnings.

Change-Id: Idf8e03f7083d5fc9ff30081651bd4ac1f2dc5bd7
Signed-off-by: Panchajanya1999 <panchajanya@azure-dev.live>